+++ src/main/java/se/swedenconnect/ca/signservice/SignserviceCaApplication.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2022 Sweden Connect
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package se.swedenconnect.ca.signservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+/**
+ * Sign service CA application main class.
+ */
+@EnableScheduling
+@SpringBootApplication(scanBasePackages = { "se.swedenconnect.ca.signservice", "se.swedenconnect.ca.service.base" })
+public class SignserviceCaApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(SignserviceCaApplication.class, args);
+  }
+
+}
