+++ src/main/java/se/swedenconnect/ca/signservice/ca/storage/data/EncryptedData.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2022 Sweden Connect
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package se.swedenconnect.ca.signservice.ca.storage.data;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+/**
+ * Representation of encrypted data.
+ */
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class EncryptedData {
+
+  private String kid;
+  private byte[] iv;
+  private byte[] ciphertext;
+
+}
