--- src/test/java/se/swedenconnect/ca/headless/SignserviceCaApplicationTests.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2022 Sweden Connect
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package se.swedenconnect.ca.headless;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.TestPropertySource;
-
-@TestPropertySource(locations = "classpath:application-test.properties")
-@SpringBootTest
-@ActiveProfiles("nodb")
-class SignserviceCaApplicationTests {
-
-  @Test
-  void contextLoads() {
-  }
-
-}
